requirements.txt
pyqt5~=5.15.3
PyOpenGL
PyOpenGL_accelerate

numpy~=1.20.1
matplotlib~=3.3.4